consistently refer bookmarks as bookmarks instead of favorites
