+++ app/builtin-pages/bookmarks.html
@@ -9,11 +9,9 @@
 <body>
   <div class="window">
     <div class="window-content">
-      <div class="pane-group">
-        <div class="pane" id="el-content"></div>
-      </div>
+      <div class="bookmarks-wrapper"></div>
     </div>
   </div>
-  <script src="beaker:builtin-pages/favorites.js"></script>
+  <script src="beaker:builtin-pages/bookmarks.js"></script>
 </body>
 </html>
